+++ ops/render_ctmpl.py
@@ -1,9 +1,15 @@
 #!/usr/bin/env python3
 import argparse
+import os
 import shutil
+import sys
 from pathlib import Path
 
-from .cli import info, success, error, shell
+from util.misc import info, success, error, shell
+
+# In case this will need to run on Windows systems
+if sys.platform.lower() == "win32":
+    os.system("color")
 
 
 def render_ctmpl(ctmpl_file: str, **kwargs) -> int:
